+++ comments-bundle/CHANGELOG.md
@@ -1,1 +1,5 @@
 # Contao comments bundle change log
+
+## DEV
+
+ * Replace `<div class="quote">` with `<blockquote>` (see contao/core#2244).