+++ comments-bundle/CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## DEV
 
+ * Append the parent ID to the form field IDs to prevent duplicate IDs (see contao/core-bundle#1493).
  * Replace `<div class="quote">` with `<blockquote>` (see contao/core#2244).