+++ comments-bundle/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Contao comments bundle change log
 
-## DEV
+## 4.6.0-RC1 (2018-07-16)
 
  * Remove subscriptions that are not activated within 24 hours (see contao/core-bundle#1512).
  * Append the parent ID to the form field IDs to prevent duplicate IDs (see contao/core-bundle#1493).